cleaned up types
