+++ components/Layout.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import Head from "next/head";
 import { useDispatch, useSelector } from "react-redux";
 import { logout } from "../store/slices/authSlice";
-import { RootState } from "../store";
+import { RootState } from "../shared/types";
 import { Box, Typography, styled } from "@mui/material";
 import { CustomButton } from "./buttons";
