+++ shared/types.ts
@@ -1,0 +1,12 @@
+import rootReducer from "../store/rootReducer";
+
+export type RootState = ReturnType<typeof rootReducer>;
+
+export type Task = {
+  id: string;
+  title: string;
+  isDone: boolean;
+};
+export type TaskState = {
+  tasks: Task[];
+};
