+++ store/index.ts
@@ -5,5 +5,4 @@ const store = configureStore({
   reducer: rootReducer,
 });
 
-export type RootState = ReturnType<typeof rootReducer>;
 export default store;
